Python-App-Development/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
redis==5.0.1
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pydantic==2.5.2
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
pandas==2.1.3
numpy==1.26.2
scikit-learn==1.3.2
mlflow==2.8.1
pyspark==3.5.0
prometheus-client==0.19.0
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
python-dotenv==1.0.0
gunicorn==21.2.0
apache-airflow==2.7.3
kubernetes==28.1.0 